feat: change header and version info
