+++ src/components/VersionInfo/VersionInfo.styles.ts
@@ -1,0 +1,11 @@
+import { styled } from '@mui/material/styles';
+import Box from '@mui/material/Box';
+
+export const InnerBox = styled(Box)({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  width: '100%',
+  gap: '20px',
+  paddingBottom: '40px',
+});
